restructuring old applications
