+++ panorama.json
@@ -11,7 +11,7 @@
     "product_vendor": "Palo Alto Networks",
     "product_name": "Panorama",
     "product_version_regex": ".*",
-    "min_phantom_version": "6.0.2",
+    "min_phantom_version": "6.0.1",
     "python_version": "3",
     "fips_compliant": false,
     "logo": "logo_panorama.svg",
@@ -49,7 +49,7 @@
             "description": "Validate the asset configuration for connectivity",
             "verbose": "This action logs into the device using a REST Api call to check the connection and credentials configured.",
             "type": "test",
-            "identifier": "test_asset_connectivity",
+            "identifier": "test_connectivity",
             "read_only": true,
             "parameters": {},
             "output": [],
